fully completed lab7

+++ semester_1/lab7_class_vector/vector/vector.cpp
@@ -1,11 +1,11 @@
 #include <vector.h>
 
-void Vector::Reserve(int newCapacity){
+void Vector::Reserve(size_t newCapacity){
     if (newCapacity <= capacity_) return;
     int* oldData = data_;
-    int oldSize = size_;
+    size_t oldSize = size_;
     data_ = new int[newCapacity];
-    for (int i = 0; i < oldSize; ++i){
+    for (size_t i = 0; i < oldSize; ++i){
         data_[i] = oldData[i];
     }
     delete[] oldData;
@@ -15,19 +15,19 @@ Vector::Vector()
 :data_(nullptr),
  size_(0),
  capacity_(0){};
-Vector::Vector(int size)
+Vector::Vector(size_t size)
 :data_(new int[size]),
  size_(size),
  capacity_(size){
-    for (int i = 0; i < size; ++i){
+    for (size_t i = 0; i < size; ++i){
         data_[i] = 0;
     }
 }
 Vector::Vector(std::initializer_list<int> list)
 :data_(nullptr),
- size_(static_cast<int>(list.size())), capacity_(static_cast<int>(list.size())){
+ size_(static_cast<size_t>(list.size())), capacity_(static_cast<size_t>(list.size())){
     data_ = new int[capacity_];
-    int index = 0;
+    size_t index = 0;
     for (int value : list){
         data_[index] = value;
         ++index;
@@ -39,7 +39,7 @@ Vector::Vector(const Vector& other)
  capacity_(other.capacity_){
     if (capacity_ > 0){
         data_ = new int[size_];
-        for (int i = 0; i < size_; ++i){
+        for (size_t i = 0; i < size_; ++i){
             data_[i] = other.data_[i];
         }
     }
@@ -52,15 +52,15 @@ Vector& Vector::operator=(const Vector& other){
     size_ = other.size_;
     capacity_ = other.capacity_;
     data_ = new int[capacity_];
-    for (int i = 0; i < other.size_; ++i){
+    for (size_t i = 0; i < other.size_; ++i){
         data_[i] = other.data_[i];
     }
     return *this;
 }
-int& Vector::operator[](int index){
+int& Vector::operator[](size_t index){
     return data_[index];
  }
-const int& Vector::operator[](int index) const {
+const int& Vector::operator[](size_t index) const {
     return data_[index];
  }
 Vector::~Vector(){
@@ -74,23 +74,23 @@ void Vector::Swap(Vector& other){
     data_ = other.data_;
     other.data_ = tmpData;
 
-    int tmpSize = size_;
+    size_t tmpSize = size_;
     size_ = other.size_;
     other.size_ = tmpSize;
 
-    int tmpCapacity = capacity_;
+    size_t tmpCapacity = capacity_;
     capacity_ = other.capacity_;
     other.capacity_ = tmpCapacity;
 }
-int Vector::Size() const{
+size_t Vector::Size() const{
     return size_;
 }
-int Vector::Capacity() const{
+size_t Vector::Capacity() const{
     return capacity_;
 }
 void Vector::PushBack(int elem){
     if (capacity_ == size_){
-        int newCapacity = (capacity_ == 0 ? 1 : 2 * capacity_);
+        size_t newCapacity = (capacity_ == 0 ? 1 : 2 * capacity_);
         Reserve(newCapacity);
     }
     data_[size_] = elem;
@@ -106,7 +106,7 @@ void Vector::Clear(){
 }
 std::ostream& operator<<(std::ostream& os,const Vector& v){
     os << '[';
-    for (int i = 0; i < v.size_; ++i){
+    for (size_t i = 0; i < v.size_; ++i){
         os << v.data_[i];
         if (i < v.size_ - 1){
             os << ", ";
