+++ semester_1/lab7_class_vector/vector/vector.h
@@ -5,26 +5,26 @@
 class Vector {
 private:
     int* data_; //указатель на дин. массив
-    int size_; //кол-во элементов
-    int capacity_; //выделенный размер
+    size_t size_; //кол-во элементов
+    size_t capacity_; //выделенный размер
 public:
     //*конструкторы
     Vector(); //по умолчанию
-    Vector(int size); //принимает размер, заполняет нулями
+    Vector(size_t size); //принимает размер, заполняет нулями
     Vector(std::initializer_list<int> list); //{1,2,3,4,5}
     Vector(const Vector& other); //копирования
     Vector& operator=(const Vector& other); //оператор присваивания копированием
     ~Vector(); //деструктор
     void Swap(Vector& other);
     //операторы индексирования const и неconst
-    const int& operator[](int index) const;
-    int& operator[](int index); //должен позволять менять
-    int Size() const;  //getSize
-    int Capacity() const;  //getCapacity
+    const int& operator[](size_t index) const;
+    int& operator[](size_t index); //должен позволять менять
+    size_t Size() const;  //getSize
+    size_t Capacity() const;  //getCapacity
     void PushBack(int elem); //добавляет элемент в конец вектора + реаллокация при необх
     void PopBack(); //удаляет последний элемент, capacity не меняет
     void Clear(); //делает контейнер пустым, capacity не меняет
-    void Reserve(int newCapacity); //задать новую capacity (больше)
+    void Reserve(size_t newCapacity); //задать новую capacity (больше)
     friend std::ostream& operator<<(std::ostream& os, const Vector& v); //оператор вывода, не член класса
 };
